+++ crates/biome_js_analyze/src/assists/correctness.rs
@@ -2,16 +2,12 @@
 
 use biome_analyze::declare_group;
 
-pub(crate) mod flip_bin_exp;
-pub(crate) mod inline_variable;
 pub(crate) mod organize_imports;
 
 declare_group! {
     pub (crate) Correctness {
         name : "correctness" ,
         rules : [
-            self :: flip_bin_exp :: FlipBinExp ,
-            self :: inline_variable :: InlineVariable ,
             self :: organize_imports :: OrganizeImports ,
         ]
      }